requirements-a2a.txt
# Real A2A Demo Dependencies

# Core FastAPI and HTTP
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2

# Data handling and validation
pydantic==2.5.0

# Async and concurrency
asyncio-mqtt==0.13.0

# Additional utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# For health checking and monitoring
requests==2.31.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1 